requirements.txt
duckduckgo_search==7.4.3
geopandas==1.0.1
#gradio
kaleido==0.2.1
langchain==0.3.21
langchain-community==0.3.20
litellm==1.63.14
llama-index==0.12.25
llama-index-llms-azure-openai==0.3.2
markdownify==0.14.1
openai==1.68.2
pandas==2.2.3
plotly==6.0.1
python-dotenv==1.0.1
rank_bm25==0.2.2
requests==2.32.3
shapely==2.0.7
smolagents==1.9.2
#streamlit==1.42.2